+++ icicle/src/hash/merkle_c_api.cpp
@@ -2,7 +2,7 @@
 #include "icicle/errors.h"
 #include "icicle/merkle/merkle_proof.h"
 #include "icicle/merkle/merkle_tree.h"
-#include "icicle/serialization.h"
+#include "icicle/merkle/merkle_proof_serializer.h"
 extern "C" {
 // Define an opaque pointer type for MerkleProof (similar to a handle in C)
 typedef icicle::MerkleProof* MerkleProofHandle;
@@ -211,39 +211,6 @@ eIcicleError icicle_merkle_proof_deserialize(MerkleProofHandle* proof, std::byte
     return eIcicleError::INVALID_POINTER;
   }
   *proof = new icicle::MerkleProof();
-  return icicle::BinarySerializer<icicle::MerkleProof>::deserialize(buffer, size, *proof);
-}
-
-// Serialize the MerkleProof object to a file
-eIcicleError icicle_merkle_proof_serialize_to_file(MerkleProofHandle proof, const char* filename, size_t filename_len)
-{
-  if (!proof) {
-    ICICLE_LOG_ERROR << "Cannot serialize a null MerkleProof instance.";
-    return eIcicleError::INVALID_POINTER;
-  }
-  if (!filename || !filename_len) {
-    ICICLE_LOG_ERROR << "Cannot serialize to a null filename.";
-    return eIcicleError::INVALID_POINTER;
-  }
-
-  std::string filename_str(filename, filename_len);
-  return icicle::BinarySerializer<icicle::MerkleProof>::serialize_to_file(std::move(filename_str), *proof);
-}
-
-// Deserialize the MerkleProof object from a file
-eIcicleError
-icicle_merkle_proof_deserialize_from_file(MerkleProofHandle* proof, const char* filename, size_t filename_len)
-{
-  if (!proof) {
-    ICICLE_LOG_ERROR << "Cannot deserialize into a null MerkleProof pointer.";
-    return eIcicleError::INVALID_POINTER;
-  }
-  if (!filename || !filename_len) {
-    ICICLE_LOG_ERROR << "Cannot deserialize from a null filename.";
-    return eIcicleError::INVALID_POINTER;
-  }
-  *proof = new icicle::MerkleProof();
-  std::string filename_str(filename, filename_len);
-  return icicle::BinarySerializer<icicle::MerkleProof>::deserialize_from_file(std::move(filename_str), *proof);
+  return icicle::BinarySerializer<icicle::MerkleProof>::deserialize(buffer, size, **proof);
 }
 } // extern "C"
