+++ icicle/tests/test_field_api.cpp
@@ -4,7 +4,8 @@
 #include "icicle/fri/fri_config.h"
 #include "icicle/fri/fri_proof.h"
 #include "icicle/fri/fri_transcript_config.h"
-#include "icicle/serialization.h"
+#include "icicle/sumcheck/sumcheck_proof_serializer.h"
+#include "icicle/fri/fri_proof_serializer.h"
 
 // Derive all ModArith tests and add ring specific tests here
 template <typename T>
@@ -848,7 +849,7 @@ TEST_F(FieldTestBase, SumcheckSingleInputProgram)
     // Deserialize proof
     SumcheckProof<scalar_t> deserialized_proof;
     ICICLE_CHECK(BinarySerializer<SumcheckProof<scalar_t>>::deserialize(
-      proof_bytes.data(), proof_bytes.size(), &deserialized_proof));
+      proof_bytes.data(), proof_bytes.size(), deserialized_proof));
 
     // Compare proofs
     uint nof_round_polynomials = sumcheck_proof.get_nof_round_polynomials();
@@ -964,7 +965,7 @@ TYPED_TEST(FieldTest, Fri)
         // Deserialize proof
         FriProof<TypeParam> deserialized_proof;
         ICICLE_CHECK(BinarySerializer<FriProof<TypeParam>>::deserialize(
-          proof_bytes.data(), proof_bytes.size(), &deserialized_proof));
+          proof_bytes.data(), proof_bytes.size(), deserialized_proof));
 
         // Compare proofs
         // Compare number of FRI rounds
