+++ wrappers/rust/icicle-core/Cargo.toml
@@ -13,8 +13,6 @@ once_cell = "1.10.0" # how to have it in dev-deps?
 rayon = "1.8.1"
 rand = "0.8" # TODO move to dev-dependencies
 serde = "1.0.219"
-serde_json = "1.0.140"
-base64 = "0.22.1"
 
 [dev-dependencies]
 criterion = "0.3"
