+++ wrappers/rust/icicle-curves/icicle-bls12-377/Cargo.toml
@@ -16,6 +16,7 @@ serde = "1.0.219"
 [dev-dependencies]
 criterion = "0.3"
 serial_test = "3.0.0"
+serde_json = "1.0.140"
 
 [build-dependencies]
 cmake = "0.1.50"
