+++ wrappers/rust/icicle-curves/icicle-bw6-761/Cargo.toml
@@ -17,6 +17,7 @@ serde = "1.0.219"
 [dev-dependencies]
 criterion = "0.3"
 serial_test = "3.0.0"
+serde_json = "1.0.140"
 
 [build-dependencies]
 cmake = "0.1.50"
