+++ wrappers/rust/icicle-fields/icicle-goldilocks/Cargo.toml
@@ -19,7 +19,7 @@ criterion = "0.3"
 serial_test = "3.0.0"
 risc0-core = "0.21.0"
 risc0-zkp = "0.21.0"
-
+serde_json = "1.0.140"
 [build-dependencies]
 cmake = "0.1.50"
