+++ wrappers/rust/icicle-fields/icicle-m31/Cargo.toml
@@ -15,6 +15,7 @@ serde = "1.0.219"
 
 [build-dependencies]
 cmake = "0.1.50"
+serde_json = "1.0.140"
 
 [features]
 cuda_backend = ["icicle-runtime/cuda_backend"]
