+++ wrappers/rust/icicle-fields/icicle-stark252/Cargo.toml
@@ -18,6 +18,7 @@ serde = "1.0.219"
 criterion = "0.3"
 serial_test = "3.0.0"
 lambdaworks-math = "0.6.0"
+serde_json = "1.0.140"
 
 [build-dependencies]
 cmake = "0.1.50"
